Add new scoring algorithm, point at Sanjee's API installation, Add URL params

+++ src/main/java/hofapp/DTO/PlayerWithScore.java
@@ -4,29 +4,29 @@ import hofapp.models.Player;
 
 public class PlayerWithScore extends Player {
 
-    private Integer score;
-    private Integer certainScore;
+    private Double score;
+    private Double certainScore;
 
-    public PlayerWithScore(Player player, int score, int certainScore) {
+    public PlayerWithScore(Player player, double score, double certainScore) {
         this.setId(player.getId());
         this.setFirstName(player.getFirstName());
         this.score = score;
         this.certainScore = certainScore;
     }
 
-    public Integer getCertainScore() {
+    public Double getCertainScore() {
         return certainScore;
     }
 
-    public void setCertainScore(Integer certainScore) {
+    public void setCertainScore(Double certainScore) {
         this.certainScore = certainScore;
     }
 
-    public Integer getScore() {
+    public Double getScore() {
         return score;
     }
 
-    public void setScore(Integer score) {
+    public void setScore(Double score) {
         this.score = score;
     }
